+++ crates/cargo-zng/src/res/built_in/shf.rs
@@ -1,13 +1,16 @@
 use super::*;
 
-const SHF_HELP: &str = r#"
-Run a bash script on the final pass
+// TODO(breaking) remove this
 
-Apart from running on final this tool behaves exactly like .zr-sh
+const SHF_HELP: &str = r#"
+DEPRECATED
+
+Use `.zr-sh.zr-f`
 "#;
 pub(super) fn shf() {
     help(SHF_HELP);
     if std::env::var(ZR_FINAL).is_ok() {
+        warn!(".zr-shf is deprecated, use .zr-sh.zr-f");
         sh();
     } else {
         println!("zng-res::on-final=");